requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.107.0
pydantic==2.5.0
python-multipart==0.0.6
nltk==3.8.1
python-dotenv==1.0.0
supabase==2.0.0
